requirements.txt
customtkinter
pillow
requests
urllib3